refactor: renamed localStorage.ts to storage.ts
